fix: renaming scrollarea to scrollbar again for telescoping
