+++ lib/components/src/ScrollBar/ScrollArea.tsx
