+++ lib/components/src/ScrollBar/stories/ScrollArea.stories.tsx
