+++ lib/components/src/ScrollBar/styled.tsx
