+++ lib/components/src/ScrollBar/utils/get-horizontal-position.ts
