+++ lib/components/src/ScrollBar/utils/get-horiztontal-values.ts
