+++ lib/components/src/ScrollBar/utils/get-state-values.ts
