+++ lib/components/src/ScrollBar/utils/get-vertical-position.ts
