+++ lib/components/src/ScrollBar/utils/get-vertical-values.ts
