+++ lib/components/src/index.ts
@@ -17,8 +17,8 @@ export { ActionBar } from './ActionBar/ActionBar';
 export { Spaced } from './spaced/Spaced';
 export { Placeholder } from './placeholder/placeholder';
 export { Zoom } from './Zoom/Zoom';
-export { ScrollArea } from './ScrollArea/ScrollArea';
-export type { ScrollAreaProps } from './ScrollArea/ScrollArea';
+export { ScrollArea } from './ScrollBar/ScrollArea';
+export type { ScrollAreaProps } from './ScrollBar/ScrollArea';
 
 // Forms
 export { Button } from './Button/Button';
